+++ CHANGELOG.md
@@ -3,6 +3,7 @@ In Development
 - Show "[empty]" for empty dist-info files (other than
   `zip-safe`/`not-zip-safe`)
 - Added a "Search Projects" box at the top of most pages
+- Use version 1.1.0 of `wheel-inspect`
 
 2018.10.17
 ----------